+++ dirent_compat.c
@@ -35,12 +35,12 @@ DIR *opendir(const char *name)
 
     if(name && name[0])
     {
-        size_t base_length = strlen(name);
+        size_t base_length = (int)strlen(name);
         const char *all = /* search pattern must end with suitable wildcard */
             strchr("/\\", name[base_length - 1]) ? "*" : "/*";
 
         if((dir = (DIR *) malloc(sizeof *dir)) != 0 &&
-           (dir->name = (char *) malloc(base_length + strlen(all) + 1)) != 0)
+           (dir->name = (char *) malloc(base_length + (int)strlen(all) + 1)) != 0)
         {
             strcat(strcpy(dir->name, name), all);
