+++ parse_time.c
@@ -208,7 +208,7 @@ format_time(char *buf, int i, int len, time_t t)
     tm = gmtime(&t);
     if(tm == NULL)
         return -1;
-    rc = strftime(buf + i, len - i, "%a, %d %b %Y %H:%M:%S GMT", tm);
+    rc = (int)strftime(buf + i, len - i, "%a, %d %b %Y %H:%M:%S GMT", tm);
     if(rc <= 0)                 /* yes, that's <= */
         return -1;
     return i + rc;
